dark-mode: use a not-too-dark color for navbar links
The links in the navigation bar are supposed to be colored dark gray
unless the current page is in the corresponding section.

In dark mode, however, it must be a bit lighter than the very dark
background.

So let's stop hard-coding that color and override it accordingly in dark
mode with a nice, pleasant light gray.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>