deploy: do open/close issues about broken links
There is specific code in git-scm.com's deploy workflow to open a ticket
whenever broken links were detected, and to close such a ticket when no
broken links were detected.

However, as per https://github.com/lycheeverse/lychee-action/pull/265
the way we checked for this was incorrect: `env.lychee_exit_code` had
was correct, until https://github.com/lycheeverse/lychee-action/pull/245
broke it by way of fixing another bug.

This was the reason why the broken link that was found and reported in
https://github.com/git/git-scm.com/actions/runs/13544529063/job/37852881418#step:3:135319
never made it into a GitHub issue, even if that had been the intention.

For the record, I worked on a fix for this broken link and opened
https://github.com/jnavila/git-manpages-l10n/pull/131 to incorporate
that fix.

What prevented this broken link from being detected before
above-mentioned deployment is the fact that before
https://github.com/git/git-scm.com/pull/1953 was merged, the
`git-scm.com` deployment used `http://` in the base URL, and hence the
`--remap` option used in the deployment workflow mapped all
http://git-scm.com links to local files, whereas now the `https://` base
URL is used and https://git-scm.com links are mapped, including the
offending one.

In any case let's fix opening/closing the "broken link" issues.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>