ci: check for broken git-scm.com links
Not all links are relative in the PR build; Some of the links originate
from Git's manual pages or from the ProGit book or its translations, and
those links are of the form "https://git-scm.com/...". Let's map those
into "file:///..." URLs so that lychee checks them even in offline mode.

This uncovers a broken link which is address by
https://github.com/jnavila/git-manpages-l10n/pull/131. For the time
being, this patch adds a work-around specifically for that issue lest
contributors' PRs fail through no fault of their own.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>