update-docs: extract the list of Git project-specific manual pages
There are a couple of "manual pages" that are really not interesting to
regular Git users because they affect how to work _with the Git
project_, not how to work _with Git_.

It is somewhat unfortunate (and a clear lack of separation of concerns)
that these are included in what Git's home page calls the "Git
reference".

Note that this has not been the case for a long time! Only with
ea0e16d02 (docs: include SubmittingPatches, 2018-04-16) it changed,
adding a single Git project-specific file to the site.

Since that dam was broken, 44bf76690 (index.rake: build
MyFirstContribution tutorial, 2019-08-14) and later db4e92007 (docs:
render 'MyFirstObjectWalk', 2020-06-07) introduced _further_ pages that
have nothing to do, really, with how to use Git.

In preparation for labeling these pages clearly as specific to the Git
project, let's extract the list of these _special_ pages to a central
location where it can be reused in the Hugo part of the site generation.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>