docs: mark Git project-specific pages clearly
This adds a callout at the top of the pages that are irrelevant to all
Git users except the ones who wish to contribute to the Git project
itself. There are only very few of them, anyway (about to be a few
more), but it is a good idea to ensure that readers are not confused
into believing that they have to follow the Git project's coding
guidelines if they want to use Git.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>