update-book2: optionally discover downloads even in `local_genbook2`
There used to be convenient download links for some of the ProGit book
translations.

In the transition to a Hugo-generated site, these links were lost for
all except the original English version, where the downloads were
hard-coded.

The reason is that in the Hugo world, we want to avoid network-heavy
roundtrips that would be necessary when using the `remote_genbook2`
function, and we use `local_genbook2` instead, where the information
about what downloadable assets there are (if any) is simply not there.

Let's add a way to discover the downloadable assets, by using Octokit,
but still use the local clone for the rest of the generation. The way to
enable this is by providing a `GITHUB_API_TOKEN` environment variable.
If that environment variable is not set, no dynamic discovery will be
performed.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>