community: use a shortcode...
Only so-called "shortcodes" are supported as Hugo constructs in the
`content/` pages. I had forgottent that when I wrote 1deace4d4 (docs:
include more Git project-specific pages, 2025-03-01).

Thankfully, the `ci` build pointed it out (specifically, the link
checker step).

Normally, I would now squash the fix into the offending commit, but
there have been automated changes in the meantime (which took a bit over
a quarter hour to run) and in the interest of some convenience for
myself (and to be able to go relax instead of pouring more time into
this topic branch than I already have), and of keeping the original
commits as made by the automated workflow run, I will just leave this
fix on top of the automated changes.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>