script/update-docs(l10n): add support for .adoc files
Since 2.49.0 the documentation files are named with the .adoc extension.
The `git-manpages-l10n` project followed suite and soon the
`git-html-l10n` tip commit (which contains the output of
`git-manpages-l10n` but is only updated in conjunction with Git
releases) will no longer have any `.txt` files, either. Adapt the script
to support that.

This recapitulates the changes made in a74b38c81 (script/update-docs:
add support for .adoc files, 2025-03-12) in the code that handles the
translated manual pages.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>