application.js: allow for relative URLs
When running Hugo with relative URLs, as is the case when a developer
runs `hugo` in a regular checkout of this repository, there is no
`https://` prefix, and therefore the current logic failed.

Fix this logic, and while at it, simplify the preceding code a bit.

Note that this logic needs further adjustment, then, for importing the
`pagefind.js` library: This `import`, if relative, is relative to the
location of `application.js` (which is not the case where
`baseURLPrefix` is used to set `href` and `src` attributes, which are
relative to the top-level directory of the site). This _happened_ to
work before because the bug was hidden by the bug where `baseURLPrefix`
was not adjusted correctly for `HUGO_RELATIVEURLS=true`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>