README: use `console` markup consistently
As early as 5d721a7ac (readme is markdown, 2012-03-08), the `README` was
written in Markdown, but its Unix shell command snippets were not marked
up as such (most likely because way back when, syntax highlighting in
`<pre>` blocks had not yet been dreamt up).

However, we are approaching the end of the first quarter of the 21st
century and it is time to go with the times.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>