css: make the header responsive
The icon to toggle between dark and light mode was added recently and
this made laying out the elements in the header hard.

A lot of items in the header have been using `position: absolute`. This
was probably caused by the absolute positioning of the search result
dropdown.

Transition most elements in the header to use flexbox to make
positioning more fluid and require less manual pixel pushing.

Signed-off-by: Toon Claes <toon@iotcl.com>