site-root: show forges logos on landing page
Replace the project and company logos by logos of forges that provide
Git hosting.

SVG images are used to properly scale on any screen size. The SVG images
are embedded directly into the HTML. This allows us to use CSS custom
properties (like `var(--logos-1)`) for the colors. Using this method the
colors can easily adapt to dark/light mode.

The SVG images are created in Inkscape. One source file `all.svg` has
all the logos. In this file placeholder colors are used. The placeholder
colors are replaced by `var(--xyz)` by the Hugo rendering pipeline. The
placeholder colors are:

* `#aaaaaa` -> `var(--logos-1)`
* `#999999` -> `var(--logos-2)`
* `#666666` -> `var(--logos-3)`
* `#333333` -> `var(--logos-4)`

These are lightest to darkest in light mode, and reverse in dark mode.