Allow testing the celebration by tapping the tagline six times
On mobile phones, it is a bit of a challenge to press the "2" key
followed by the "0" key to test out the 20th anniversary celebrations.
Implement a multi-tap listener for that, specifically on the "tagline"
so as not to spoil the surprise by mistake.

I considered requiring 20 taps because it is the 20th anniversary, but I
also want to test efficiently, so there.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>