content/about/trademark.html: update trademark policy
Git's existing trademark policy forbids using the Mark as a syllable in
another word or portmanteau "without Conservancy's written permission".
It similarly forbids creating and/or selling merchandise "without
Conservancy's expression written permission".

The quoted portions of the Policy have no legal effect, because a
trademark owner may always give permission regardless of whatever policy
might be in place. But they have the unintended side-effect of
encouraging others to write to Conservancy and ask for an exception to
one and/or the other, though such an exception is rarely (if ever)
granted.

Remove that portion of the trademark policy, and add a new one
explaining that GitHub and GitLab are historical exceptions to the
policy.

Acked-by: Junio C Hamano <gitster@pobox.com>
Acked-by: Christian Couder <christian.couder@gmail.com>
Signed-off-by: Taylor Blau <me@ttaylorr.com>