dark-mode: allow following the system preference again
After toggling dark mode on and off again, a subsequent change of the
system preference to dark mode will not respected because the toggling
stored the user's preference.

Let's store the user preference only when it differs from the user's
choice. That way, in the scenario above the system preference would be
followed again.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>